pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pwc-agentic-doc-processor"
version = "1.0.0"
description = "AI-Powered Agentic Document Processing System"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Citta AI", email = "team@citta.ai"},
]
keywords = ["ai", "document-processing", "vertex-ai", "gemini", "langgraph", "rag"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "streamlit>=1.30.0",
    "google-genai>=1.0.0",
    "google-cloud-aiplatform>=1.29.0",
    "langgraph>=0.2.0",
    "PyPDF2>=3.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[project.urls]
Repository = "https://github.com/citta-ai/pwc-agentic-doc-processor"
Documentation = "https://github.com/citta-ai/pwc-agentic-doc-processor#readme"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
